Move email related code to utilities.email.
